Update BmiCalculator.cs
123

+++ WebBMI/HealthMgr/BmiCalculator.cs
@@ -20,8 +20,8 @@ namespace HealthMgr
         public float Calculate()
         {
             float result = 0;
-            //將 float 改成 int 試試看 XD
-            float height = (float)Height / 100;
+            //將 float 改成 int 試試看
+            float height = (int)Height / 100;
             result = Weight / (height * height);
 
             return result;
